agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph>=1.1,<2
langgraph-prebuilt>=1.0,<2
langchain-core>=1.3,<2
langchain-mcp-adapters>=0.1,<1
httpx==0.27.2
python-dotenv==1.0.1
pydantic>=2.9,<3
pydantic-settings>=2.5,<3
apscheduler==3.10.4
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
